Entrenadores ya pueden consultar calificaciones

+++ proyecto1/src/main/java/com/unam/proyecto1/servicio/CalificacionServicio.java
@@ -3,5 +3,5 @@ package com.unam.proyecto1.servicio;
 import com.unam.proyecto1.modelo.Calificacion;
 
 public interface CalificacionServicio {
-    Calificacion creaCalificacion(String nombreEvento, String nombreDisciplina, String categoria, String rama, String emailJuez, String emailCompetidor, int puntaje, String comentario);
+    Calificacion creaCalificacion(String nombreEvento, String nombreDisciplina, String categoria, String rama, String emailJuez, String emailCompetidor, double puntaje, String comentario);
 }
